Bumps gnutls26 version to gnutls28.
This script pulls, patches and compiles a version of gnutls that
is no longer available on debian or Ubuntu, as per
https://tracker.debian.org/news/582966.

That makes the script fail, unless an old version of Debian is used.
Unfortunately, old debian means old GHC and/or old dependencies, which
make some newer packages break more often.

This commit bumps the version of gnutls to that available in more recent
versions of Ubuntu/Debian, but at the expense of no longer supporting
older versions that did not provide gnutls28.